+++ otx-pool/src/logo.rs
