+++ otx-pool/src/plugin_extension/host_service.rs
