+++ src/main.rs
@@ -1,6 +1,6 @@
 use atomic_swap::AtomicSwap;
 use dust_collector::DustCollector;
-use otx_pool::{cli::print_logo, OtxPoolService};
+use otx_pool::{logo::print_logo, OtxPoolService};
 use signer::Signer;
 use utils::config::{parse, AppConfig, ConfigFile};
